cast size_t to System::UInt32 in 64 bit

+++ AbstractionLayer.DotNET/AL/DotNET/GPIO/UARTDevice.hpp
@@ -115,7 +115,7 @@ namespace AL::DotNET::GPIO
 				buffer,
 				&_buffer[0],
 				offset,
-				_buffer.GetSize()
+				static_cast<System::UInt32>(_buffer.GetSize())
 			);
 
 			try
